+++ lib/features/email/presentation/widgets/bottom_bar_mail_widget_builder.dart
