+++ lib/features/mailbox_dashboard/presentation/mailbox_dashboard_view.dart
@@ -2,7 +2,7 @@ import 'package:core/core.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:jmap_dart_client/jmap/core/session/session.dart';
-import 'package:tmail_ui_user/features/mail/presentation/mail_view.dart';
+import 'package:tmail_ui_user/features/email/presentation/email_view.dart';
 import 'package:tmail_ui_user/features/mailbox/presentation/mailbox_view.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/mailbox_dashboard_controller.dart';
 import 'package:tmail_ui_user/features/thread/presentation/thread_view.dart';
@@ -33,7 +33,7 @@ class MailboxDashBoardView extends GetWidget<MailboxDashBoardController> {
               flex: responsiveUtils.isDesktop(context) ? 1 : 3,
               child: ThreadView(),
             ),
-            if (responsiveUtils.isDesktop(context)) Expanded(flex: 2, child: MailView()),
+            if (responsiveUtils.isDesktop(context)) Expanded(flex: 2, child: EmailView()),
           ],
         ),
       ),
