+++ lib/main/pages/app_pages.dart
@@ -3,8 +3,8 @@ import 'package:tmail_ui_user/features/home/presentation/home_bindings.dart';
 import 'package:tmail_ui_user/features/home/presentation/home_view.dart';
 import 'package:tmail_ui_user/features/login/presentation/login_bindings.dart';
 import 'package:tmail_ui_user/features/login/presentation/login_view.dart';
-import 'package:tmail_ui_user/features/mail/presentation/mail_bindings.dart';
-import 'package:tmail_ui_user/features/mail/presentation/mail_view.dart';
+import 'package:tmail_ui_user/features/email/presentation/email_bindings.dart';
+import 'package:tmail_ui_user/features/email/presentation/email_view.dart';
 import 'package:tmail_ui_user/features/mailbox/presentation/mailbox_bindings.dart';
 import 'package:tmail_ui_user/features/mailbox/presentation/mailbox_view.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/mailbox_dashboard_bindings.dart';
@@ -42,8 +42,8 @@ class AppPages {
       page: () => ThreadView(),
       binding: ThreadBindings()),
     GetPage(
-      name: AppRoutes.MAIL,
-      page: () => MailView(),
-      binding: MailBindings()),
+      name: AppRoutes.EMAIL,
+      page: () => EmailView(),
+      binding: EmailBindings()),
   ];
 }
